requirements.txt
discord~=1.7.3
googletrans~=4.0.0rc1
aiohttp~=3.7.4.post0
asyncio~=3.4.3
requests~=2.27.1
pyfiglet~=0.8.post1
Pillow~=8.4.0
python-dotenv~=0.19.2